+++ src/main/scala/org/podval/tools/scalajsplugin/scalajs/ScalaJSRunTask.scala
@@ -1,6 +1,0 @@
-package org.podval.tools.scalajsplugin.scalajs
-
-import org.podval.tools.node.Node
-import org.podval.tools.scalajsplugin.nonjvm.NonJvmRunTask
-
-trait ScalaJSRunTask extends NonJvmRunTask[ScalaJSLinkTask] with ScalaJSTask