+++ src/main/scala/org/podval/tools/scalaplugin/scalajs/ScalaJSLinkMainTask.scala
@@ -1,0 +1,13 @@
+package org.podval.tools.scalaplugin.scalajs
+
+import org.gradle.api.NamedDomainObjectContainer
+import org.gradle.api.tasks.Nested
+import org.podval.tools.build.scalajs.ModuleInitializer
+import org.podval.tools.scalaplugin.nonjvm.NonJvmLinkTask
+import scala.jdk.CollectionConverters.SetHasAsScala
+
+abstract class ScalaJSLinkMainTask extends NonJvmLinkTask.Main[ScalaJSLinkTask] with ScalaJSLinkTask:
+  @Nested def getModuleInitializers: NamedDomainObjectContainer[ModuleInitializerProperties]
+
+  final override def moduleInitializers: Option[Seq[ModuleInitializer]] =
+    Some(getModuleInitializers.asScala.toSeq.map(_.toModuleInitializer))