+++ src/main/scala/org/podval/tools/scalaplugin/scalanative/ScalaNativeTask.scala
@@ -1,0 +1,5 @@
+package org.podval.tools.scalaplugin.scalanative
+
+import org.podval.tools.scalaplugin.nonjvm.NonJvmTask
+
+trait ScalaNativeTask extends NonJvmTask[ScalaNativeLinkTask]