+++ src/main/scala/org/podval/tools/scalaplugin/scalanative/ScalaNativeTestTask.scala
@@ -1,0 +1,13 @@
+package org.podval.tools.scalaplugin.scalanative
+
+import org.podval.tools.build.scalanative.{ScalaNativeBuild, ScalaNativeTestEnvironment}
+import org.podval.tools.scalaplugin.nonjvm.NonJvmRunTask
+
+abstract class ScalaNativeTestTask extends NonJvmRunTask.Test[ScalaNativeLinkTask] with ScalaNativeRunTask:
+  final override protected def linkTaskClass: Class[ScalaNativeLinkTestTask] = classOf[ScalaNativeLinkTestTask]
+  
+  final override protected def createTestEnvironment: ScalaNativeTestEnvironment =
+    ScalaNativeBuild.createTestEnvironment(
+      binaryTestFile = linkTask.getOutputFile,
+      logSource = getName
+    )