+++ src/main/scala/org/podval/tools/test/task/TestTask.scala
@@ -12,7 +12,6 @@ import org.gradle.internal.time.Clock
 import org.gradle.internal.work.WorkerLeaseService
 import org.gradle.internal.{Actions, Cast}
 import org.gradle.util.internal.ConfigureUtil
-import org.podval.tools.build.ScalaBackend
 import org.podval.tools.test.environment.TestEnvironment
 import java.io.File
 import java.lang.reflect.Method
@@ -26,13 +25,6 @@ object TestTask:
     useTestFramework.invoke(task, value)
 
 abstract class TestTask extends Test:
-  // TODO I'd love to remove this and use getTestEnvironment.backend instead,
-  // but since the test task is materialized when replacing the pre-existing one,
-  // `useSbt` is called on it, which calls `createTestEnvironment` before the classpath is
-  // extended with the backend-specific stuff...
-  // Look into calling `useSbt` via a convention - if Gradle does this when creating its own test task!
-  protected def backend: ScalaBackend
-
   protected def createTestEnvironment: TestEnvironment
   
   final def useSbt(@DelegatesTo(classOf[SbtTestFrameworkOptions]) testFrameworkConfigure: Closure[?]): Unit =
@@ -56,7 +48,6 @@ abstract class TestTask extends Test:
       testEnvironment = None
 
   private lazy val sbtTestFramework: SbtTestFramework = SbtTestFramework(
-    backend = backend,
     logLevelEnabled = getServices.get(classOf[StartParameter]).getLogLevel,
     defaultTestFilter = getFilter.asInstanceOf[DefaultTestFilter],
     options = SbtTestFrameworkOptions(),
@@ -64,6 +55,7 @@ abstract class TestTask extends Test:
     testTaskTemporaryDir = getTemporaryDirFactory,
     dryRun = getDryRun,
     // delayed: not available at the time of the TestFramework construction (task creation)
+    backend = () => getTestEnvironment.backend,
     loadedFrameworks = (testClassPath: Iterable[File]) => getTestEnvironment.loadedFrameworks(testClassPath)
   )
   
@@ -82,7 +74,7 @@ abstract class TestTask extends Test:
       closeTestEnvironment()
 
   final override def createTestExecuter: TestExecuter = TestExecuter(
-    testsCanNotBeForked = backend.testsCanNotBeForked,
+    testsCanNotBeForked = getTestEnvironment.backend.testsCanNotBeForked,
     sourceMapper = getTestEnvironment.sourceMapper,
     testFilter = getFilter.asInstanceOf[DefaultTestFilter],
     maxWorkerCount = getServices.get(classOf[StartParameter]).getMaxWorkerCount,
@@ -95,6 +87,6 @@ abstract class TestTask extends Test:
   )
 
   final override def getMaxParallelForks: Int = 
-    if backend.testsCanNotBeForked
+    if getTestEnvironment.backend.testsCanNotBeForked
     then 1
     else super.getMaxParallelForks